Crate ethers_contract_abigen

Abigen

Programmatically generate type-safe Rust bindings for Ethereum smart contracts.

This crate is intended to be used either indirectly with the abigen procedural macro or directly from a build script / CLI.

Re-exports

pub use filter::ContractFilter;

pub use filter::ExcludeContracts;

pub use filter::SelectContracts;

pub use multi::MultiAbigen;

Modules

contract

Contains types to generate Rust bindings for Solidity contracts.

filter

Filtering support for contracts used in Abigen

multi

Generate bindings for multiple Abigen

Structs

Abigen

Programmatically generate type-safe Rust bindings for an Ethereum smart contract from its ABI.

ContractBindings

Type-safe contract bindings generated by Abigen.

InternalStructs

Helper to match ethabi::Params with structs and nested structs

Enums

Source

A source of an Ethereum smart contract’s ABI.

Type Definitions

Address