Crate ethers_contract_abigen

Module for generating type-safe bindings to Ethereum smart contracts. This module is intended to be used either indirectly with the abigen procedural macro or directly from a build script / CLI

Modules

contract

Contains types to generate rust bindings for solidity contracts

rawabi

This is a basic representation of a contract ABI that does no post processing but contains the raw content of the ABI.

Structs

Abigen

Builder struct for generating type-safe bindings from a contract’s ABI

ContractBindings

Type-safe contract bindings generated by a Builder. This type can be either written to file or into a token stream for use in a procedural macro.

MultiAbigen

Generates bindings for a series of contracts

Enums

Source

A source of a Truffle artifact JSON.

Functions

parse_address

Parses the given address string

Type Definitions

Address