Struct ethers_contract_abigen::contract::Context

pub struct Context { /* private fields */ }

Internal shared context for generating smart contract bindings.

Implementations

impl Context

pub fn events_declaration(&self) -> Result<TokenStream>

Expands each event to a struct + its impl Detokenize block

pub fn event_methods(&self) -> Result<TokenStream>

Generate the event filter methods for the contract

impl Context

Expands a context into a method struct containing all the generated bindings to the Solidity contract methods.

pub fn expand_return_struct( &self, function: &Function, alias: Option<&MethodAlias> ) -> Result<Option<TokenStream>>

Expands to the corresponding struct type based on the inputs of the given function

impl Context

pub fn abi_structs(&self) -> Result<TokenStream>

Generate corresponding types for structs parsed from a human readable ABI

NOTE: This assumes that all structs that are potentially used as type for variable are in fact present in the AbiParser, this is sound because AbiParser::parse would have failed already

pub fn remove_struct(&mut self, name: &str)

In the event of type conflicts this allows for removing a specific struct type.

pub fn struct_definition(&mut self, name: &str) -> Result<TokenStream>

Returns the type definition for the struct with the given name

impl Context

pub fn expand(&self) -> Result<ExpandedContract>

Generates the tokens.

pub fn from_abigen(args: Abigen) -> Result<Self>

Create a context from the code generation arguments.

pub fn internal_structs(&self) -> &InternalStructs

Returns a reference to the internal ABI struct mapping table.

pub fn internal_structs_mut(&mut self) -> &mut InternalStructs

Returns a mutable reference to the internal ABI struct mapping table.